+++ audionimbus/src/effect/ambisonics/rotation.rs
@@ -1,11 +1,13 @@
 use super::super::AudioEffectState;
+use super::super::EffectError;
 use crate::audio_buffer::{AudioBuffer, Sample};
 use crate::audio_settings::AudioSettings;
 use crate::context::Context;
 use crate::error::{to_option_error, SteamAudioError};
 use crate::ffi_wrapper::FFIWrapper;
 use crate::geometry::CoordinateSystem;
-use crate::ChannelPointers;
+use crate::num_ambisonics_channels;
+use crate::{ChannelPointers, ChannelRequirement};
 
 /// Applies a rotation to an ambisonics audio buffer.
 ///
@@ -50,7 +52,13 @@ use crate::ChannelPointers;
 /// # Ok::<(), Box<dyn std::error::Error>>(())
 /// ```
 #[derive(Debug)]
-pub struct AmbisonicsRotationEffect(audionimbus_sys::IPLAmbisonicsRotationEffect);
+pub struct AmbisonicsRotationEffect {
+    inner: audionimbus_sys::IPLAmbisonicsRotationEffect,
+
+    /// The number of input and output channels needed for the ambisonics order used when creating
+    /// the effect.
+    num_channels: u32,
+}
 
 impl AmbisonicsRotationEffect {
     /// Creates a new ambisonics rotation effect.
@@ -63,7 +71,7 @@ impl AmbisonicsRotationEffect {
         audio_settings: &AudioSettings,
         ambisonics_rotation_effect_settings: &AmbisonicsRotationEffectSettings,
     ) -> Result<Self, SteamAudioError> {
-        let mut ambisonics_rotation_effect = Self(std::ptr::null_mut());
+        let mut inner = std::ptr::null_mut();
 
         let status = unsafe {
             audionimbus_sys::iplAmbisonicsRotationEffectCreate(
@@ -72,7 +80,7 @@ impl AmbisonicsRotationEffect {
                 &mut audionimbus_sys::IPLAmbisonicsRotationEffectSettings::from(
                     ambisonics_rotation_effect_settings,
                 ),
-                ambisonics_rotation_effect.raw_ptr_mut(),
+                &mut inner,
             )
         };
 
@@ -80,39 +88,54 @@ impl AmbisonicsRotationEffect {
             return Err(error);
         }
 
+        let num_channels = num_ambisonics_channels(ambisonics_rotation_effect_settings.max_order);
+        let ambisonics_rotation_effect = Self {
+            inner,
+            num_channels,
+        };
+
         Ok(ambisonics_rotation_effect)
     }
 
     /// Applies an ambisonics rotation effect to an audio buffer.
     ///
     /// This effect CANNOT be applied in-place.
+    ///
+    /// Both input and output audio buffers must have as many channels as needed for the
+    /// Ambisonics order used (see [`crate::num_ambisonics_channels`]).
+    ///
+    /// # Errors
+    ///
+    /// Returns [`EffectError`] if:
+    /// - The input buffer does not have the correct number of channels for the Ambisonics order
+    /// - The output buffer does not have the correct number of channels for the Ambisonics order
     pub fn apply<I, O, PI: ChannelPointers, PO: ChannelPointers>(
         &mut self,
         ambisonics_rotation_effect_params: &AmbisonicsRotationEffectParams,
         input_buffer: &AudioBuffer<I, PI>,
         output_buffer: &AudioBuffer<O, PO>,
-    ) -> AudioEffectState
+    ) -> Result<AudioEffectState, EffectError>
     where
         I: AsRef<[Sample]>,
         O: AsRef<[Sample]> + AsMut<[Sample]>,
     {
-        let required_num_channels = (ambisonics_rotation_effect_params.order + 1).pow(2);
-        assert_eq!(
-            input_buffer.num_channels(),
-            required_num_channels,
-            "ambisonic order N = {} requires (N + 1)^2 = {} input channels",
-            ambisonics_rotation_effect_params.order,
-            required_num_channels
-        );
-        assert_eq!(
-            output_buffer.num_channels(),
-            required_num_channels,
-            "ambisonic order N = {} requires (N + 1)^2 = {} output channels",
-            ambisonics_rotation_effect_params.order,
-            required_num_channels
-        );
+        let num_input_channels = input_buffer.num_channels();
+        if num_input_channels != self.num_channels {
+            return Err(EffectError::InvalidInputChannels {
+                expected: ChannelRequirement::Exactly(self.num_channels),
+                actual: num_input_channels,
+            });
+        }
 
-        unsafe {
+        let num_output_channels = output_buffer.num_channels();
+        if num_output_channels != self.num_channels {
+            return Err(EffectError::InvalidOutputChannels {
+                expected: ChannelRequirement::Exactly(self.num_channels),
+                actual: num_output_channels,
+            });
+        }
+
+        let state = unsafe {
             audionimbus_sys::iplAmbisonicsRotationEffectApply(
                 self.raw_ptr(),
                 &mut *ambisonics_rotation_effect_params.as_ffi(),
@@ -120,7 +143,9 @@ impl AmbisonicsRotationEffect {
                 &mut *output_buffer.as_ffi(),
             )
         }
-        .into()
+        .into();
+
+        Ok(state)
     }
 
     /// Retrieves a single frame of tail samples from an Ambisonics rotation effect’s internal buffers.
@@ -128,17 +153,32 @@ impl AmbisonicsRotationEffect {
     /// After the input to the Ambisonics rotation effect has stopped, this function must be called instead of [`Self::apply`] until the return value indicates that no more tail samples remain.
     ///
     /// The output audio buffer must have as many channels as needed for the Ambisonics order specified when creating the effect.
-    pub fn tail<O>(&self, output_buffer: &AudioBuffer<O>) -> AudioEffectState
+    ///
+    /// # Errors
+    ///
+    /// Returns [`EffectError`] if the output buffer does not have the correct number of channels
+    /// for the Ambisonics order.
+    pub fn tail<O>(&self, output_buffer: &AudioBuffer<O>) -> Result<AudioEffectState, EffectError>
     where
         O: AsRef<[Sample]> + AsMut<[Sample]>,
     {
-        unsafe {
+        let num_output_channels = output_buffer.num_channels();
+        if num_output_channels != self.num_channels {
+            return Err(EffectError::InvalidOutputChannels {
+                expected: ChannelRequirement::Exactly(self.num_channels),
+                actual: num_output_channels,
+            });
+        }
+
+        let state = unsafe {
             audionimbus_sys::iplAmbisonicsRotationEffectGetTail(
                 self.raw_ptr(),
                 &mut *output_buffer.as_ffi(),
             )
         }
-        .into()
+        .into();
+
+        Ok(state)
     }
 
     /// Returns the number of tail samples remaining in an Ambisonics rotation effect’s internal buffers.
@@ -157,29 +197,33 @@ impl AmbisonicsRotationEffect {
     ///
     /// This is intended for internal use and advanced scenarios.
     pub fn raw_ptr(&self) -> audionimbus_sys::IPLAmbisonicsRotationEffect {
-        self.0
+        self.inner
     }
 
     /// Returns a mutable reference to the raw FFI pointer.
     ///
     /// This is intended for internal use and advanced scenarios.
     pub fn raw_ptr_mut(&mut self) -> &mut audionimbus_sys::IPLAmbisonicsRotationEffect {
-        &mut self.0
+        &mut self.inner
     }
 }
 
 impl Clone for AmbisonicsRotationEffect {
     fn clone(&self) -> Self {
         unsafe {
-            audionimbus_sys::iplAmbisonicsRotationEffectRetain(self.0);
+            audionimbus_sys::iplAmbisonicsRotationEffectRetain(self.inner);
         }
-        Self(self.0)
+
+        Self {
+            inner: self.inner,
+            num_channels: self.num_channels,
+        }
     }
 }
 
 impl Drop for AmbisonicsRotationEffect {
     fn drop(&mut self) {
-        unsafe { audionimbus_sys::iplAmbisonicsRotationEffectRelease(&mut self.0) }
+        unsafe { audionimbus_sys::iplAmbisonicsRotationEffectRelease(&mut self.inner) }
     }
 }
 
@@ -226,5 +270,183 @@ impl AmbisonicsRotationEffectParams {
             };
 
         FFIWrapper::new(ambisonics_rotation_effect_params)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::*;
+
+    mod apply {
+        use super::*;
+
+        #[test]
+        fn test_valid_first_order() {
+            let context = Context::default();
+            let audio_settings = AudioSettings::default();
+
+            let mut effect = AmbisonicsRotationEffect::try_new(
+                &context,
+                &audio_settings,
+                &AmbisonicsRotationEffectSettings { max_order: 1 },
+            )
+            .unwrap();
+
+            let params = AmbisonicsRotationEffectParams {
+                orientation: CoordinateSystem::default(),
+                order: 1,
+            };
+
+            let input = vec![0.5; 4 * 1024];
+            let input_buffer = AudioBuffer::try_with_data_and_settings(
+                &input,
+                AudioBufferSettings::with_num_channels(4),
+            )
+            .unwrap();
+
+            let mut output = vec![0.0; 4 * 1024];
+            let output_buffer = AudioBuffer::try_with_data_and_settings(
+                &mut output,
+                AudioBufferSettings::with_num_channels(4),
+            )
+            .unwrap();
+
+            assert!(effect.apply(&params, &input_buffer, &output_buffer).is_ok());
+        }
+
+        #[test]
+        fn test_invalid_input_channels() {
+            let context = Context::default();
+            let audio_settings = AudioSettings::default();
+
+            let mut effect = AmbisonicsRotationEffect::try_new(
+                &context,
+                &audio_settings,
+                &AmbisonicsRotationEffectSettings { max_order: 1 },
+            )
+            .unwrap();
+
+            let params = AmbisonicsRotationEffectParams {
+                orientation: CoordinateSystem::default(),
+                order: 1,
+            };
+
+            let input = vec![0.5; 2 * 1024];
+            let input_buffer = AudioBuffer::try_with_data_and_settings(
+                &input,
+                AudioBufferSettings::with_num_channels(2),
+            )
+            .unwrap();
+
+            let mut output = vec![0.0; 4 * 1024];
+            let output_buffer = AudioBuffer::try_with_data_and_settings(
+                &mut output,
+                AudioBufferSettings::with_num_channels(4),
+            )
+            .unwrap();
+
+            assert_eq!(
+                effect.apply(&params, &input_buffer, &output_buffer),
+                Err(EffectError::InvalidInputChannels {
+                    expected: ChannelRequirement::Exactly(4),
+                    actual: 2
+                })
+            );
+        }
+
+        #[test]
+        fn test_invalid_output_channels() {
+            let context = Context::default();
+            let audio_settings = AudioSettings::default();
+
+            let mut effect = AmbisonicsRotationEffect::try_new(
+                &context,
+                &audio_settings,
+                &AmbisonicsRotationEffectSettings { max_order: 1 },
+            )
+            .unwrap();
+
+            let params = AmbisonicsRotationEffectParams {
+                orientation: CoordinateSystem::default(),
+                order: 1,
+            };
+
+            let input = vec![0.5; 4 * 1024];
+            let input_buffer = AudioBuffer::try_with_data_and_settings(
+                &input,
+                AudioBufferSettings::with_num_channels(4),
+            )
+            .unwrap();
+
+            let mut output = vec![0.0; 3 * 1024];
+            let output_buffer = AudioBuffer::try_with_data_and_settings(
+                &mut output,
+                AudioBufferSettings::with_num_channels(3),
+            )
+            .unwrap();
+
+            assert_eq!(
+                effect.apply(&params, &input_buffer, &output_buffer),
+                Err(EffectError::InvalidOutputChannels {
+                    expected: ChannelRequirement::Exactly(4),
+                    actual: 3,
+                })
+            );
+        }
+    }
+
+    mod tail {
+        use super::*;
+
+        #[test]
+        fn test_valid() {
+            let context = Context::default();
+            let audio_settings = AudioSettings::default();
+
+            let effect = AmbisonicsRotationEffect::try_new(
+                &context,
+                &audio_settings,
+                &AmbisonicsRotationEffectSettings { max_order: 1 },
+            )
+            .unwrap();
+
+            let mut output = vec![0.0; 4 * 1024];
+            let output_buffer = AudioBuffer::try_with_data_and_settings(
+                &mut output,
+                AudioBufferSettings::with_num_channels(4),
+            )
+            .unwrap();
+
+            assert!(effect.tail(&output_buffer).is_ok());
+        }
+
+        #[test]
+        fn test_invalid_output_channels() {
+            let context = Context::default();
+            let audio_settings = AudioSettings::default();
+
+            let effect = AmbisonicsRotationEffect::try_new(
+                &context,
+                &audio_settings,
+                &AmbisonicsRotationEffectSettings { max_order: 1 },
+            )
+            .unwrap();
+
+            let mut output = vec![0.0; 2 * 1024];
+            let output_buffer = AudioBuffer::try_with_data_and_settings(
+                &mut output,
+                AudioBufferSettings::with_num_channels(2),
+            )
+            .unwrap();
+
+            assert_eq!(
+                effect.tail(&output_buffer),
+                Err(EffectError::InvalidOutputChannels {
+                    expected: ChannelRequirement::Exactly(4),
+                    actual: 2,
+                })
+            );
+        }
     }
 }